Move OfferHeader to ExpandableCardContent
