+++ src/components/PlaceOfferSection/PlaceOfferSection.module.less
@@ -1,33 +1,3 @@
-.content {
-  padding: 12px 24px 24px;
-  cursor: default;
-  width: 100%;
-
-  @media (max-width: 640px) {
-    padding: 16px;
-  }
-}
-
-.offerHeaderContent {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-
-  @media (max-width: 640px) {
-    margin-bottom: 0;
-  }
-}
-.offerHeaderTitle {
-  font: var(--title-sm);
-  text-transform: uppercase;
-  line-height: 38px;
-
-  @media (max-width: 640px) {
-    font: var(--title-xs);
-    line-height: 32px;
-  }
-}
-
 .borrowerMessage {
   font: var(--body-text-sm);
 }
