+++ src/components/PlaceOfferSection/PlaceOfferSection.tsx
@@ -1,11 +1,7 @@
 import { FC } from 'react'
 
 import PlaceOfferContent from './PlaceOfferContent'
-import { OfferHeader } from './components'
-import { checkIsEditMode } from './helpers'
 import { usePlaceOffer } from './hooks'
-
-import styles from './PlaceOfferSection.module.less'
 
 interface PlaceOfferSectionProps {
   offerPubkey: string
@@ -13,27 +9,10 @@ interface PlaceOfferSectionProps {
   setOfferPubkey?: (offerPubkey: string) => void
 }
 
-const PlaceOfferSection: FC<PlaceOfferSectionProps> = ({
-  offerPubkey,
-  marketPubkey,
-  setOfferPubkey,
-}) => {
-  const offerParams = usePlaceOffer({
-    offerPubkey,
-    marketPubkey,
-    setOfferPubkey,
-  })
+const PlaceOfferSection: FC<PlaceOfferSectionProps> = (props) => {
+  const offerParams = usePlaceOffer(props)
 
-  const { exitEditMode } = offerParams
-
-  return (
-    <div className={styles.content}>
-      {setOfferPubkey && (
-        <OfferHeader isEditMode={checkIsEditMode(offerPubkey)} exitEditMode={exitEditMode} />
-      )}
-      <PlaceOfferContent {...offerParams} />
-    </div>
-  )
+  return <PlaceOfferContent {...offerParams} />
 }
 
 export default PlaceOfferSection
