+++ src/components/PlaceOfferSection/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useOfferTransactions'
 export * from './usePlaceOffer'
 export * from './useLenderLoans'
+export * from './useSyntheticOffer'
