+++ src/pages/LendPage/components/ExpandableCardContent/ExpandableCardContent.module.less
@@ -6,7 +6,7 @@
   width: 100%;
 }
 
-.content {
+.tabsContent {
   border: var(--border-less-primary);
   border-right: 0;
   border-bottom: 0;
@@ -24,5 +24,41 @@
 
   .ant-modal-content {
     padding: 0;
+  }
+}
+
+.content {
+  padding: 12px 24px 24px;
+  cursor: default;
+  width: 100%;
+
+  @media (max-width: 640px) {
+    padding: 16px;
+  }
+}
+
+.offerHeaderContent {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+
+  @media (max-width: 640px) {
+    margin-bottom: 0;
+  }
+}
+.offerHeaderTitle {
+  font: var(--title-sm);
+  text-transform: uppercase;
+  line-height: 38px;
+
+  @media (max-width: 640px) {
+    font: var(--title-xs);
+    line-height: 32px;
+  }
+}
+
+.offersButton {
+  @media (min-width: 1241px) {
+    display: none;
   }
 }
