+++ src/pages/LendPage/components/ExpandableCardContent/ExpandableCardContent.tsx
@@ -1,14 +1,13 @@
 import { FC, useState } from 'react'
 
-import PlaceOfferSection from '@banx/components/PlaceOfferSection/PlaceOfferSection'
-import { Tab, Tabs, useTabs } from '@banx/components/Tabs'
-import { Modal } from '@banx/components/modals/BaseModal'
+import PlaceOfferSection, {
+  checkIsEditMode,
+  useSyntheticOffer,
+} from '@banx/components/PlaceOfferSection'
 
 import { useModal } from '@banx/store'
-import { toLowerCaseNoSpaces, trackPageEvent } from '@banx/utils'
 
-import ActivityTable from '../ActivityTable'
-import OrderBook from '../OrderBook'
+import { OfferHeader, OffersModal, TabsContent } from './components'
 
 import styles from './ExpandableCardContent.module.less'
 
@@ -18,21 +17,34 @@ interface ExpandableCardContentProps {
 
 const ExpandableCardContent: FC<ExpandableCardContentProps> = ({ marketPubkey }) => {
   const [offerPubkey, setOfferPubkey] = useState('')
-  // const { open } = useModal()
+  const { open } = useModal()
 
-  // const showModal = () => {
-  //   open(OffersModal, { setOfferPubkey, offerPubkey, marketPubkey })
-  // }
+  const { removeSyntheticOffer } = useSyntheticOffer(offerPubkey, marketPubkey)
+
+  const exitEditMode = () => {
+    setOfferPubkey('')
+    removeSyntheticOffer()
+  }
+
+  const showModal = () => {
+    open(OffersModal, { setOfferPubkey, offerPubkey, marketPubkey })
+  }
 
   return (
     <div className={styles.container}>
-      <PlaceOfferSection
-        offerPubkey={offerPubkey}
-        marketPubkey={marketPubkey}
-        setOfferPubkey={setOfferPubkey}
-      />
-
       <div className={styles.content}>
+        <OfferHeader
+          isEditMode={checkIsEditMode(offerPubkey)}
+          showModal={showModal}
+          exitEditMode={exitEditMode}
+        />
+        <PlaceOfferSection
+          offerPubkey={offerPubkey}
+          marketPubkey={marketPubkey}
+          setOfferPubkey={setOfferPubkey}
+        />
+      </div>
+      <div className={styles.tabsContent}>
         <TabsContent
           marketPubkey={marketPubkey}
           offerPubkey={offerPubkey}
@@ -44,60 +56,3 @@ const ExpandableCardContent: FC<ExpandableCardContentProps> = ({ marketPubkey })
 }
 
 export default ExpandableCardContent
-
-interface MarketParams {
-  marketPubkey: string
-  offerPubkey: string
-  setOfferPubkey: (offerPubkey: string) => void
-}
-
-const TabsContent = ({ marketPubkey, offerPubkey, setOfferPubkey }: MarketParams) => {
-  const { value: currentTabValue, ...tabsProps } = useTabs({
-    tabs: BONDS_TABS,
-    defaultValue: BONDS_TABS[0].value,
-  })
-
-  const onTabClick = (tabProps: Tab) => {
-    trackPageEvent('lend', `${toLowerCaseNoSpaces(tabProps.label)}tab`)
-  }
-
-  return (
-    <>
-      <Tabs value={currentTabValue} onTabClick={onTabClick} {...tabsProps} />
-      {currentTabValue === TabName.OFFERS && (
-        <OrderBook
-          marketPubkey={marketPubkey}
-          offerPubkey={offerPubkey}
-          setOfferPubkey={setOfferPubkey}
-        />
-      )}
-      {currentTabValue === TabName.ACTIVITY && <ActivityTable marketPubkey={marketPubkey} />}
-    </>
-  )
-}
-
-export const OffersModal: FC<MarketParams> = (props) => {
-  const { close } = useModal()
-
-  return (
-    <Modal className={styles.modal} open onCancel={close}>
-      <TabsContent {...props} />
-    </Modal>
-  )
-}
-
-enum TabName {
-  OFFERS = 'offers',
-  ACTIVITY = 'activity',
-}
-
-const BONDS_TABS = [
-  {
-    label: 'Offers',
-    value: TabName.OFFERS,
-  },
-  {
-    label: 'Activity',
-    value: TabName.ACTIVITY,
-  },
-]
